Added update timeout

+++ ARI/update.sh
@@ -11,10 +11,15 @@ if [[ "$currentver" == "$newver" ]]
 then
 	echo "You have the latest version ($currentver) installed!"
 else
-	echo "A new update available ($newver)! Downloading..."
-	git clone https://github.com/Marakusa/ari.git /tmp/ari/ari
-	cd /tmp/ari/ari
-	sudo chmod a+x install.sh
-	bash ./install.sh
-	sudo rm -rf /tmp/ari/ari
+	if [[ "$newver" == "" ]]
+	then
+		echo "Timed out. Please check your internet connection."
+	else
+		echo "A new update available ($newver)! Downloading..."
+		git clone https://github.com/Marakusa/ari.git /tmp/ari/ari
+		cd /tmp/ari/ari
+		sudo chmod a+x install.sh
+		bash ./install.sh
+		sudo rm -rf /tmp/ari/ari
+	fi
 fi
